[FEATURE] Adding the regex that you were looking for
I erase the complex string replacement that you were doing and added the regex that you were looking for.
You work is very interesting and impressive. I was looking how to download on Pornhub and your code give me the trick. And you'r right, that's gross Haha !

Anyway, you should learn the lookahead and lookbehind for the regex, that's a powerful tool.